--- runbot_pylint/model/runbot_pylint_pylint.txt
@@ -1,49 +1,0 @@
-************* Module openerp.addons.runbot_pylint
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/__init__.py:1: [C0111(missing-docstring), ] Missing module docstring
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/__init__.py:27: [W0403(relative-import), ] Relative import 'model', should be 'openerp.addons.runbot_pylint.model'
-************* Module openerp.addons.runbot_pylint.model.runbot_inhe
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:40: [C0301(line-too-long), ] Line too long (82/80)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:40: [C0326(bad-whitespace), ] No space allowed around keyword argument assignment
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:45: [C0303(trailing-whitespace), ] Trailing whitespace
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:48: [C0301(line-too-long), ] Line too long (87/80)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:55: [C0326(bad-whitespace), ] Exactly one space required around assignment
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:56: [C0301(line-too-long), ] Line too long (94/80)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:1: [C0111(missing-docstring), ] Missing module docstring
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:33: [C0103(invalid-name), ] Invalid constant name "logger"
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:35: [C0103(invalid-name), runbot_repo] Invalid class name "runbot_repo"
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:35: [C0111(missing-docstring), runbot_repo] Missing class docstring
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:35: [R0904(too-many-public-methods), runbot_repo] Too many public methods (66/20)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:43: [C0103(invalid-name), runbot_build] Invalid class name "runbot_build"
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:43: [C0111(missing-docstring), runbot_build] Missing class docstring
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:46: [R0913(too-many-arguments), runbot_build.job_30_run] Too many arguments (7/5)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:47: [W0621(redefined-outer-name), runbot_build.job_30_run] Redefining name 'logger' from outer scope (line 33)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:46: [C0103(invalid-name), runbot_build.job_30_run] Invalid argument name "cr"
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:46: [C0111(missing-docstring), runbot_build.job_30_run] Missing method docstring
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:56: [W0212(protected-access), runbot_build.job_30_run] Access to a protected member _run_test_pylint of a client class
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:56: [E1101(no-member), runbot_build.job_30_run] Instance of 'runbot_build' has no 'pool' member
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:46: [W0613(unused-argument), runbot_build.job_30_run] Unused argument 'args'
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:43: [R0904(too-many-public-methods), runbot_build] Too many public methods (67/20)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:27: [W0611(unused-import), ] Unused import argparse
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:31: [W0611(unused-import), ] Unused import os
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:30: [W0611(unused-import), ] Unused import subprocess
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/runbot_inhe.py:32: [W0611(unused-import), ] Unused import tools
-************* Module openerp.addons.runbot_pylint.model
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/__init__.py:1: [C0111(missing-docstring), ] Missing module docstring
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/__init__.py:27: [W0403(relative-import), ] Relative import 'runbot_inhe', should be 'openerp.addons.runbot_pylint.model.runbot_inhe'
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/__init__.py:28: [W0403(relative-import), ] Relative import 'pylint', should be 'openerp.addons.runbot_pylint.model.pylint'
-************* Module openerp.addons.runbot_pylint.model.pylint
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:34: [C0103(invalid-name), ] Invalid constant name "logger"
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:50: [C0103(invalid-name), PylintConf._run_test_pylint] Invalid argument name "cr"
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:61: [W1201(logging-not-lazy), PylintConf._run_test_pylint] Specify string format arguments as logging function parameters
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:50: [W0613(unused-argument), PylintConf._run_test_pylint] Unused argument 'uid'
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:50: [W0613(unused-argument), PylintConf._run_test_pylint] Unused argument 'cr'
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:64: [R0201(no-self-use), PylintConf.run_command] Method could be a function
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:36: [R0904(too-many-public-methods), PylintConf] Too many public methods (67/20)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:107: [R0913(too-many-arguments), PylintError.name_search] Too many arguments (8/5)
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:107: [W0221(arguments-differ), PylintError.name_search] Arguments number differs from overridden method
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:145: [W0221(arguments-differ), PylintError.name_get] Arguments number differs from overridden method
-/home/ernesto/instancias/odoo/odoo_original/openerp/addons/runbot_pylint/model/pylint.py:95: [R0904(too-many-public-methods), PylintError] Too many public methods (66/20)
-|openerp.addons.runbot_pylint.model.runbot_inhe              |0.14  |0.18    |0.20     |0.10       |
-|openerp.addons.runbot_pylint.model.pylint                   |0.00  |0.13    |0.26     |0.01       |
-|openerp.addons.runbot_pylint.model.__init__                 |0.00  |0.05    |0.00     |0.01       |
-|openerp.addons.runbot_pylint.__init__                       |0.00  |0.03    |0.00     |0.01       |
